small tweaks

+++ code/modules/1713/complex_animal/_backend.dm
@@ -14,6 +14,8 @@
 	icon_state = dead_state
 
 /mob/living/simple_animal/complex_animal/proc/taming(var/mob/living/carbon/human/H, var/value = 0)
+	if (!can_be_tamed)
+		return
 	if (!tamed)
 		var/found = FALSE
 		if (value != 0)
@@ -29,6 +31,8 @@
 	else
 		return
 /mob/living/simple_animal/complex_animal/proc/check_tame_stat(var/mob/living/carbon/human/H)
+	if (!can_be_tamed)
+		return
 	if (friendly_mobs[H.name] && friendly_mobs[H.name][1] == H)
 		if (friendly_mobs[H.name][2] >= tameminimum && !owner && !tamed)
 			owner = H
@@ -47,24 +51,25 @@
 	onTouchedBy(H, H.a_intent)
 
 /mob/living/simple_animal/complex_animal/attackby(var/obj/item/weapon/W as obj, var/mob/living/carbon/human/H as mob)
-	if (H.a_intent == I_HELP)
-		if (istype(W, /obj/item/weapon/reagent_containers/food))
-			if (istype(W, /obj/item/weapon/reagent_containers/food/snacks/meat))
-				var/obj/item/weapon/reagent_containers/food/snacks/meat/MT = W
-				if (MT.rotten)
-					H << "The meat is rotten! \The [src] doesn't want it..."
-					taming(H, -1)
+	if (can_be_tamed)
+		if (H.a_intent == I_HELP)
+			if (istype(W, /obj/item/weapon/reagent_containers/food))
+				if (istype(W, /obj/item/weapon/reagent_containers/food/snacks/meat))
+					var/obj/item/weapon/reagent_containers/food/snacks/meat/MT = W
+					if (MT.rotten)
+						H << "The meat is rotten! \The [src] doesn't want it..."
+						taming(H, -1)
+						return
+					else if (MT.reagents.has_reagent("protein",1))
+						nutrition += 18*MT.reagents.get_reagent_amount("protein")
+						H << "\The [src] eats \the [MT] and wags it's tail in happiness!"
+						taming(H, 5)
+						H.drop_from_inventory(W)
+						qdel(W)
+						return
+				else
+					H << "\The [src] sniffs the [W] but does not eat it."
 					return
-				else if (MT.reagents.has_reagent("protein",1))
-					nutrition += 18*MT.reagents.get_reagent_amount("protein")
-					H << "\The [src] eats \the [MT] and wags it's tail in happiness!"
-					taming(H, 5)
-					H.drop_from_inventory(W)
-					qdel(W)
-					return
-			else
-				H << "\The [src] sniffs the [W] but does not eat it."
-				return
 	..(W, H)
 	onAttackedBy(H, W)
